requirements.txt
PyYAML
python-dotenv
SQLAlchemy
psycopg[binary,pool]
alembic
langchain
langchain-community
langchain-openai
langchain-google-vertexai
langchain_google_genai
pypdf
python-docx
unstructured
tqdm
tenacity
fastapi
uvicorn[standard]
croniter
pgvector
boto3
python-multipart